added nodemailer and app verification Routes
Now when an appointment is made it has initial active status false. The customer receives and email with a confirmation link. When the link is clicked if the appointment with the id in the url exists it will set active status to true and the customer will receive a confirmation email